+++ src/lisp/kernel/lsp/packages.lisp
@@ -174,7 +174,6 @@
             sequence-stream
             all-encodings
             make-encoding
-            assume-right-type
             assert-error
             float-nan-p
             float-infinity-p
@@ -225,6 +224,8 @@
             tpl-frame tpl-argument tpl-arguments
             ;; GC
             garbage-collect finalize save-lisp-and-die
-            ;; Misc
+            ;; Compiler
+            describe-compiler-policy
             with-current-source-form
+            ;; Misc
             printing-char-p)))