+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Version 2.2.0 (LLVM14) Pending
 
+## Added
+* `ext:describe-compiler-policy` to get information about the current
+  behavior of the compiler.
+
 ## Fixes
 * Ensure that `print-unreadable-object` can accept output stream designators.
 * Set stream column to zero after printing the prompt in a REPL. Fixes [#1348][].